feat: endpoint para buscar conta por id | exception para credenciais invalidas

+++ backend/bank/src/main/java/com/avaliacaods/bank/repositories/AccountsRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.avaliacaods.bank.models.Conta;
 
 @Repository
-public interface AccountsRepository extends JpaRepository<Conta,Integer>{
+public interface AccountsRepository extends JpaRepository<Conta,Long>{
     List<Conta> findByClienteUserId(Long userId);
 
     boolean existsByNumero(String numero);
